Struct connection_pool::ConnectionPoolBehaviour

pub struct ConnectionPoolBehaviour { /* private fields */ }

Implementations

impl ConnectionPoolBehaviour

pub fn dial(&mut self, address: Multiaddr, out: Sender<Option<Contact>>)

Dial address, and send contact back on success None means something prevented us from connecting - dial reach failure or something else

pub fn connect(&mut self, new_contact: Contact, outlet: Sender<bool>)

Connect to the contact by all of its known addresses and return whether connection succeeded If contact is already being dialed and there are no new addresses in Contact, don’t dial If contact is already connected, return true immediately

pub fn disconnect(&mut self, peer_id: PeerId, outlet: Sender<bool>)

pub fn is_connected(&self, peer_id: PeerId, outlet: Sender<bool>)

Returns whether given peer is connected or not

pub fn get_contact(&self, peer_id: PeerId, outlet: Sender<Option<Contact>>)

Returns contact for a given peer if it is known

pub fn send( &mut self, to: Contact, particle: ExtendedParticle, outlet: Sender<SendStatus> )

Sends a particle to a connected contact. Returns whether sending succeeded or not Result is sent to channel inside upgrade_outbound in ProtocolHandler

pub fn count_connections(&mut self, outlet: Sender<usize>)

Returns number of connected contacts

pub fn add_subscriber(&mut self, outlet: UnboundedSender<LifecycleEvent>)

Subscribes given channel for all LifecycleEvents

pub fn add_discovered_addresses( &mut self, peer_id: PeerId, addresses: Vec<Multiaddr> )

impl ConnectionPoolBehaviour

pub fn new( buffer: usize, protocol_config: ProtocolConfig, peer_id: PeerId, metrics: Option<ConnectionPoolMetrics> ) -> (Self, Receiver<ExtendedParticle>, ConnectionPoolApi)